components/screens: remove import cycle
The AvalancheForecast component is now only a screen, and we had an
import cycle between it and the forecast screens. This commit moves all
the screen-related components to the screens directory to break the
cycle.
